Add exception handling for MBean disappearing between listing and evaluation
There was a condition where evaluating the set of added MBean can cause
an exception. This would occur if the MBean had disappeared between
listing the MBean and extracting the MBean info. If this happened in the
middle of a set of added MBeans, it would mean our current MBeans set in
the JMXHarvester isn't up to date with the current MBeans.

This has cascading impact down the line because there are MBeans
registered from previous runs that are not correctly associated with the
set of current MBeans. This means they will never be cleaned up. This
has impact on the collect stage which then results in other exceptions.

We found this occur during Repair jobs when the repair durations are
very small.